Added `helpers/pane.rs` as a generic Feathers pane component.
